% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{extdata/HG001_GRCh38.table}
\alias{extdata/HG001_GRCh38.table}
\title{Dataset HG001_GRCh38}
\format{
A data frame with 4786095 rows and 12 variables:
\describe{
\item{CHROM}{Chromosome}
\item{POS}{Position}
\item{REF}{Referenece}
\item{ALT}{Alternative}
\item{FILTER}{VQSR pass}
\item{QUAL}{Quality}
\item{QD}{QualByDepth}
\item{FS}{FisherStrand}
\item{SOR}{StrandOddsRatio}
\item{MQ}{RMSMappingQuality}
\item{MQRankSum}{MappingQualityRankSumTest}
\item{ReadPosRankSum}{ReadPosRankSumTest}
}
}
\source{
\url{https://www.nist.gov/programs-projects/genome-bottle}
}
\description{
A dataset containing GATK quality annotations for a VCF file
generated using GATK Germline short variant discovery pipeline on
HG001 reference sample from genome in a bottle project.
This dataset is used as an example to display this package's functions
}
\section{HG001_GRCh38.table}{
NA
}

