% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FS.R
\name{FS}
\alias{FS}
\title{Plot FisherStrand (FS)}
\usage{
FS(f, n)
}
\arguments{
\item{f}{VCF input file}

\item{n}{Threshold value (optional), default is the recommended GATK value.}
}
\value{
A plot of QD score and frequency
with table summary
and number of variants falling outside the threshold
}
\description{
This is the Phred-scaled probability that there is strand bias at the site. Strand Bias tells us whether the alternate allele was seen more or less often on the forward or reverse strand than the reference allele. When there little to no strand bias at the site, the FS value will be close to 0
}
