% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQ.R
\name{MQ}
\alias{MQ}
\title{Plot Mapping Quality (MQ)}
\usage{
MQ(f, n)
}
\arguments{
\item{f}{VCF input file}

\item{n}{Threshold value (optional), default is the recommended GATK value.}
}
\value{
A plot of MQ score and frequency
with table summary
and number of variants falling outside the threshold
}
\description{
This is the root mean square mapping quality over all the reads at the site. Instead of the average mapping quality of the site, this annotation gives the square root of the average of the squares of the mapping qualities at the site. It is meant to include the standard deviation of the mapping qualities. Including the standard deviation allows us to include the variation in the dataset. A low standard deviation means the values are all close to the mean, whereas a high standard deviation means the values are all far from the mean.When the mapping qualities are good at a site, the MQ will be around 60.
}
