% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QD.R
\name{QD}
\alias{QD}
\title{Plot Quality by Depth (QD)}
\usage{
QD(f, n)
}
\arguments{
\item{f}{VCF input file}

\item{n}{Threshold value (optional), default is the recommended GATK value.}
}
\value{
A plot of QD score and frequency
with table summary
and number of variants falling outside the threshold
}
\description{
This is the variant confidence (from the QUAL field) divided by the unfiltered depth of non-hom-ref samples. This annotation is intended to normalize the variant quality in order to avoid inflation caused when there is deep coverage. For filtering purposes it is better to use QD than either QUAL or DP directly. The generic filtering recommendation for QD is to filter out variants with QD below 2.
}
