% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadPosRankSum.R
\name{ReadPosRankSum}
\alias{ReadPosRankSum}
\title{Plot ReadPosRankSumTest (ReadPosRankSum)}
\usage{
ReadPosRankSum(f, n)
}
\arguments{
\item{f}{VCF input file}

\item{n}{Threshold value (optional), default is the recommended GATK value.}
}
\value{
A plot of ReadPosRankSum score and frequency
with table summary
and number of variants falling outside the threshold
}
\description{
This is the u-based z-approximation from the Rank Sum Test for site position within reads. It compares whether the positions of the reference and alternate alleles are different within the reads. Seeing an allele only near the ends of reads is indicative of error, because that is where sequencers tend to make the most errors. A negative value indicates that the alternate allele is found at the ends of reads more often than the reference allele; a positive value indicates that the reference allele is found at the ends of reads more often than the alternate allele. A value close to zero is best because it indicates there is little difference between the positions of the reference and alternate alleles in the reads.
}
