% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOR.R
\name{SOR}
\alias{SOR}
\title{Plot StrandOddsRAtio (SOR)}
\usage{
SOR(f, n)
}
\arguments{
\item{f}{file}

\item{n}{threshold}
}
\value{
A plot of SOR score and frequency
with table summary
and number of variants falling outside the threshold
}
\description{
This is another way to estimate strand bias using a test similar to the symmetric odds ratio test. SOR was created because FS tends to penalize variants that occur at the ends of exons. Reads at the ends of exons tend to only be covered by reads in one direction and FS gives those variants a bad score. SOR will take into account the ratios of reads that cover both alleles.
}
